+++ libs/client/features/src/user-details/index.ts
@@ -1,1 +1,2 @@
 export * from './UserDetails'
+export * from './UserDevTools'
